refactor: style dashboard columns
