--- src/components/DropArea/index.ts
@@ -1,1 +1,0 @@
-export * from "./DropArea";
