+++ src/components/index.ts
@@ -1,4 +1,3 @@
 export * from "./AuthWrapper";
 export * from "./WorkflowWrapper";
 export * from "./PriorityTag";
-export * from "./DropArea";
