+++ src/features/Workflow/components/DashboardColumn/style.css
@@ -1,7 +1,28 @@
 .dashboard-column {
   width: 100%;
+  border: 1px solid #d9d9d9;
+  background-color: #f0f0f0;
+  border-radius: 6px;
+  height: 100%;
+  min-height: calc(100vh - (180px + 146px));
+}
+
+.dashboard-column .ant-card-body {
+  padding: 16px;
 }
 
 .ticket-card {
   box-shadow: 0px 2px 4px 0px #00000014;
+}
+
+.drop-area {
+  border-radius: 4px;
+  opacity: 1;
+  background-color: rgba(0, 0, 0, 0.4);
+  transition: all 0.3s;
+}
+
+.hide-drop-area {
+  background-color: #f0f0f0;
+  transition: all 0.3s;
 }
